# Customer-Order-Portal/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart On X Customer Ordering Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portal/location-requests:
    get:
      operationId: listPortalLocationRequests
      tags: [portal]
      summary: List location requests by current user's org
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationRequest"
    post:
      operationId: createLocationRequest
      tags: [portal]
      summary: Submit a new location request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PORTAL MANAGE TEAM ──────────────────────────────────────────
...
  /portal/manage/locations:
    get:
      operationId: listManageLocations
      tags: [portal]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HsLocation"
    post:
      operationId: createManageLocation
      tags: [portal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
  /portal/manage/locations/{id}:
    patch:
      operationId: updateManageLocation
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
...
  /portal/manage/location-requests/{id}/approve:
    post:
      operationId: approveLocationRequest
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveLocationRequestInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /portal/manage/location-requests/{id}/reject:
    post:
      operationId: rejectLocationRequest
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectLocationRequestInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── PORTAL MANAGE BILLING ──────────────────────────────────────
...
  /portal/manage/billing/{id}:
    patch:
      operationId: updateManageBilling
      tags: [portal]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsBillingProfile"
  /portal/manage/billing/new:
    post:
      operationId: createManageBilling
      tags: [portal]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillingProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsBillingProfile"

  # ─── ADMIN DASHBOARD ─────────────────────────────────────────────
...
  /admin/organizations/{id}/locations:
    get:
      operationId: listAdminOrgLocations
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HsLocation"
    post:
      operationId: createAdminOrgLocation
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
  /admin/organizations/{id}/locations/{locId}:
    patch:
      operationId: updateAdminOrgLocation
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: locId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HsLocation"
...
  /admin/location-requests:
    get:
      operationId: listAdminLocationRequests
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationRequest"
  /admin/location-requests/{id}/approve:
    post:
      operationId: adminApproveLocationRequest
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveLocationRequestInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/location-requests/{id}/reject:
    post:
      operationId: adminRejectLocationRequest
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectLocationRequestInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── ADMIN INTEGRATION LOG ──────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ─────────────────────────────────────────────
...
    HsBillingProfile:
      type: object
      required: [id, companyId, name, address1, city, state, zip, apEmail, paymentTerms, isDefault, validated]
      properties:
        id:
          type: string
        companyId:
          type: string
        name:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        apEmail:
          type: string
        apPhone:
          type: ["string", "null"]
        paymentTerms:
          type: string
        netDays:
          type: ["integer", "null"]
        isDefault:
          type: boolean
        validated:
          type: boolean
        placeId:
          type: ["string", "null"]
        usedByLocationCount:
          type: number
    HsLocation:
      type: object
      required: [id, companyId, billingProfileId, nickname, address1, city, state, zip, isDefault, validated]
      properties:
        id:
          type: string
        companyId:
          type: string
        billingProfileId:
          type: string
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        isDefault:
          type: boolean
        validated:
          type: boolean
        placeId:
          type: ["string", "null"]
        validationSource:
          type: string
        billingProfile:
          $ref: "#/components/schemas/HsBillingProfile"
...
    LocationRequest:
      type: object
      required: [id, hubspotCompanyId, requestedByUserId, nickname, address1, city, state, zip, status, createdAt]
      properties:
        id:
          type: number
        hubspotCompanyId:
          type: string
        requestedByUserId:
          type: number
        requestedByUsername:
          type: ["string", "null"]
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        validated:
          type: boolean
        placeId:
          type: ["string", "null"]
        status:
          type: string
        resolvedBy:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        orgName:
          type: ["string", "null"]
    LocationRequestInput:
      type: object
      required: [nickname, address1, city, state, zip]
      properties:
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
          minLength: 2
          maxLength: 2
        zip:
          type: string
        phone:
          type: ["string", "null"]
        placeId:
          type: ["string", "null"]
        validated:
          type: boolean
    ApproveLocationRequestInput:
      type: object
      required: [billingProfileId]
      properties:
        billingProfileId:
          type: string
        isDefault:
          type: boolean
    RejectLocationRequestInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ─── TEAM SCHEMAS ─────────────────────────────────────────────
...
    LocationInput:
      type: object
      required: [nickname, address1, city, state, zip, billingProfileId]
      properties:
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        billingProfileId:
          type: string
        isDefault:
          type: boolean
        placeId:
          type: ["string", "null"]
        validated:
          type: boolean
    LocationUpdate:
      type: object
      properties:
        nickname:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        phone:
          type: ["string", "null"]
        billingProfileId:
          type: string
        isDefault:
          type: boolean
        placeId:
          type: ["string", "null"]
        validated:
          type: boolean

    # ─── BILLING SCHEMAS ─────────────────────────────────────────
...
    BillingProfileInput:
      type: object
      required: [name, address1, city, state, zip, apEmail, paymentTerms]
      properties:
        name:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        apEmail:
          type: string
        apPhone:
          type: ["string", "null"]
        paymentTerms:
          type: string
        netDays:
          type: ["integer", "null"]
        isDefault:
          type: boolean
        placeId:
          type: ["string", "null"]
    BillingProfileUpdate:
      type: object
      properties:
        name:
          type: string
        address1:
          type: string
        address2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        apEmail:
          type: string
        apPhone:
          type: ["string", "null"]
        paymentTerms:
          type: string
        netDays:
          type: ["integer", "null"]
        isDefault:
          type: boolean

    # ─── ADMIN ORG SCHEMAS ────────────────────────────────────────
...